further refactoring and install perfect freehand
