+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@douyinfe/semi-ui": "^2.66.1",
     "canvas": "^2.11.2",
+    "perfect-freehand": "^1.2.2",
     "react": "^18.3.1",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
