+++ src/utils/canvasConstants.ts
